Refactor Telemetry  - Change Namespace  - Add static Names class for Meter and Activity names
