+++ CarrotMQ.Core.Test/MessageProcessing/MessageDistributorTest.cs
@@ -8,8 +8,8 @@ using CarrotMQ.Core.MessageProcessing.Delivery;
 using CarrotMQ.Core.MessageProcessing.Middleware;
 using CarrotMQ.Core.Protocol;
 using CarrotMQ.Core.Serialization;
+using CarrotMQ.Core.Telemetry;
 using CarrotMQ.Core.Test.Helper;
-using CarrotMQ.Core.Tracing;
 using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
