+++ CarrotMQ.Core/Telemetry/Names.cs
@@ -1,0 +1,14 @@
+﻿namespace CarrotMQ.Core.Telemetry;
+
+///
+public static class Names
+{
+    /// <summary>
+    /// Name of the meter used for CarrotMQ metrics.
+    /// </summary>
+    public const string CarrotMeterName = "CarrotMQ.Meter";
+    /// <summary>
+    /// Name of the ActivitySource object used for tracing with CarrotMQ
+    /// </summary>
+    public const string CarrotActivitySourceName = "CarrotMQ.ActivitySource";
+}
